Add thread

+++ word/word_view.py
@@ -4,13 +4,14 @@ import wx.lib.agw.thumbnailctrl as tc
 from urllib import FancyURLopener
 import urllib2
 import simplejson
-import time
-from random import shuffle
+from random import shuffle, randint
 import os.path
 from dictionary.database import ImageDB
 from gui.utils import convert_image_to_string, convert_string_to_image
 import shutil
+import thread
 GOOGLE_IMAGE_SERVER = 'https://encrypted-tbn0.gstatic.com/images?q=tbn:'
+IMAGE_PATH_FORMAT = '%s/%s.jpg'
 
 
 class WordDisplay(wx.Dialog):
@@ -19,16 +20,16 @@ class WordDisplay(wx.Dialog):
     """
     def __init__(self, parent, title, content):
         super(WordDisplay, self).__init__(parent, wx.ID_ANY, 'Definition of ' + title)
+        self.is_running = True
         self.word_images_directory = 'image'
-        self.title = title
-        self.image_db = ImageDB()
-        self.image_dict = self.image_db.load()
-        self.images_saved = self.image_dict.get(self.title.lower(), [])
-        self.load_from_db_to_image()
-        self.content = content
-        # self.word_images_directory = 'image/%s' % self.title.lower()
         if not os.path.exists(self.word_images_directory):
             os.makedirs(self.word_images_directory)
+        self.title = title.lower()
+        self.image_db = ImageDB()
+        self.image_dict = self.image_db.load()
+        self.images_saved = self.image_dict.get(self.title, [])
+        self.load_from_db_to_image()
+        self.content = content
         self.thumbnails = tc.ThumbnailCtrl(self, -1, thumboutline=tc.THUMB_OUTLINE_FULL,
                                            imagehandler=tc.NativeImageHandler)
         main_sizer = self.design_interface()
@@ -36,11 +37,15 @@ class WordDisplay(wx.Dialog):
         self.Bind(wx.EVT_CLOSE, self.on_close)
 
     def load_from_db_to_image(self):
+        """
+        Load string from database to save to images
+        :return:
+        """
         image_count = 0
         for image_str in self.images_saved:
             image_count += 1
             image = convert_string_to_image(image_str)
-            image_path = '%s/%s.jpg' % (self.word_images_directory, str(image_count))
+            image_path = IMAGE_PATH_FORMAT % (self.word_images_directory, str(image_count))
             image.save(image_path)
 
     def design_interface(self):
@@ -62,7 +67,7 @@ class WordDisplay(wx.Dialog):
         Create word title on top of dialog
         :return:
         """
-        word_title = wx.StaticText(self, wx.ID_ANY, self.title, style=wx.ALIGN_CENTER_HORIZONTAL)
+        word_title = wx.StaticText(self, wx.ID_ANY, self.title.upper(), style=wx.ALIGN_CENTER_HORIZONTAL)
         word_title.SetForegroundColour(wx.Colour(255, 0, 0))
         font = wx.Font(18, wx.ROMAN, wx.NORMAL, wx.BOLD)
         word_title.SetFont(font)
@@ -75,7 +80,7 @@ class WordDisplay(wx.Dialog):
         """
         description_image_sizer = wx.BoxSizer(wx.VERTICAL)
         self.thumbnails.ShowFileNames(False)
-        self.thumbnails.ShowDir(self.word_images_directory)
+        self.show_image_thumbs()
         change_image_btn = wx.Button(self, -1, 'Get randomly word description images from Google Image Search')
         change_image_btn.Bind(wx.EVT_BUTTON, self.change_description_images)
         description_image_sizer.Add(self.thumbnails, 1, wx.EXPAND | wx.ALL, 10)
@@ -88,35 +93,51 @@ class WordDisplay(wx.Dialog):
         :param e:
         :return:
         """
-        nb_images = 0
         self.images_saved = []
-        # get first images from google search result
-        rnd_numbers = [i for i in range(0, 10)]
-        shuffle(rnd_numbers)
-        for i in range(0, len(rnd_numbers)):
-            # each request from encrypt server return a list of 4 images
-            url = ('https://ajax.googleapis.com/ajax/services/search/images?' + 'v=1.0&q=' + self.title +
-                   '&start=' + str(rnd_numbers[i] * 4) + '&userip=MyIP')
-            request = urllib2.Request(url)
-            response = urllib2.urlopen(request)
+        # select images in 20 first result images from google search
+        rnd_number = randint(0, 20)
+        url = ('https://ajax.googleapis.com/ajax/services/search/images?' + 'v=1.0&q=' + self.title +
+               '&start=' + str(rnd_number) + '&userip=MyIP')
+        request = urllib2.Request(url)
+        response = urllib2.urlopen(request)
+        results = simplejson.load(response)
+        data = results['responseData']
+        data_info = data['results']
+        opener = FancyURLopener()
+        shuffle(data_info)
+        self.is_running = True
+        thread.start_new_thread(self.thread_image, (data_info, opener))
 
-            results = simplejson.load(response)
-            data = results['responseData']
-            data_info = data['results']
-
-            opener = FancyURLopener()
-            shuffle(data_info)
-            for u in data_info:
-                nb_images += 1
-                image_path = '%s/%s.jpg' % (self.word_images_directory, str(nb_images))
+    def thread_image(self, data_info, opener):
+        """
+        A thread to handle the processing searching images from google
+        :param data_info:
+        :param opener:
+        :return:
+        """
+        if not self.is_running:
+            self.is_running = False
+            thread.exit()
+            return
+        count = 0
+        for u in data_info:
+            count += 1
+            try:
+                if not os.path.exists(self.word_images_directory):
+                    os.makedirs(self.word_images_directory)
+                image_path = IMAGE_PATH_FORMAT % (self.word_images_directory, str(count))
                 # key 'imageId' to get ID of image (Note: key 'unescapedUrl' for origin url of image)
                 opener.retrieve(GOOGLE_IMAGE_SERVER + u['imageId'], image_path)
                 self.images_saved.append(convert_image_to_string(image_path))
-                if nb_images == 3:
-                    self.thumbnails.ShowDir(self.word_images_directory)
-                    return
-            # avoid to be blocked IP by google (this code can not be reached if data_info has more than 3 images)
-            time.sleep(1)
+            except IOError:
+                print 'Please wait some seconds before fetching another images.'
+            if len(self.images_saved) == 3:
+                self.show_image_thumbs()
+                self.is_running = False
+                thread.exit()
+                return
+        self.is_running = False
+        thread.exit()
 
     def create_definition_box(self):
         """
@@ -131,8 +152,21 @@ class WordDisplay(wx.Dialog):
         return definition_box_sizer
 
     def on_close(self, e):
-        self.image_dict[self.title.lower()] = self.images_saved
+        """
+        Event raises when the dialog is closed
+        :param e:
+        :return:
+        """
+        self.image_dict[self.title] = self.images_saved
         self.image_db.save(self.image_dict)
-        shutil.rmtree(self.word_images_directory)
-        os.makedirs(self.word_images_directory)
         self.Destroy()
+
+    def show_image_thumbs(self):
+        """
+        Display thumbnails of image from a directory
+        Note: ThumbnailCtrl does not support the viewing image from paths
+        :return:
+        """
+        if os.path.exists(self.word_images_directory):
+            self.thumbnails.ShowDir(self.word_images_directory)
+            shutil.rmtree(self.word_images_directory)
